add: move and register row modules.
